requirements.txt
fastapi
uvicorn
python-dotenv
pypdf
python-multipart

faiss-cpu
openai
tiktoken

sentence-transformers
streamlit
requests